requirements.txt
requests>=2.18.0
requests_oauthlib>=1.2.0
python-dateutil>=2.7
pytz>=2018.5
tzlocal>=1.5.0
beautifulsoup4>=4.0.0
stringcase>=1.2.0
pycrypto==2.6.1
